pyproject.toml
[tool.poetry]
name = "vscodeextensionsmanager"
version = "0.1.0"
description = ""
authors = ["faronan"]

[tool.poetry.dependencies]
python = "^3.10"
pytablewriter = "^0.64.1"
bs4 = "^0.0.1"
requests = "^2.26.0"
isort = "^5.10.1"
black = "^21.12b0"
pyproject-flake8 = "^0.0.1-alpha.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203,"
max-complexity = 10

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
